Deploy preview on any non master branch

+++ travis.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 SCRIPT_BASE="$( cd -P "$( dirname "$0" )" && pwd )"
-if [[ -d ${SCRIPT_BASE}/public ]]; then
+if [[ -d ${SCRIPT_BASE}/public && "${TRAVIS_COMMIT_MESSAGE}" == *"clean"* ]]; then
     rm -rf ${SCRIPT_BASE}/public
 fi
 DRAFTS=""
-if [[ "$TRAVIS_BRANCH" = "preview" ]]; then
-    DRAFTS="-D -b http://blogpreview.erikzaadi.com/"
+if [[ "$TRAVIS_BRANCH" != "master" ]]; then
+    DRAFTS="-D -F -b http://blogpreview.erikzaadi.com/"
 fi
 ${SCRIPT_BASE}/bin/hugo ${DRAFTS}
